feat(Validator): 시도 횟수는 양의 숫자만 가능하다

+++ src/main/java/racingcar/service/RaceService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class RaceService {
 
-    private int numOfStages;
+    private long numOfStages;
     private List<Car> carList;
 
     public void raceStart() {
@@ -23,7 +23,7 @@ public class RaceService {
         this.carList = InputParser.getCarList(userInputDTO);
     }
 
-    public int getNumOfStages() {
+    public long getNumOfStages() {
         return this.numOfStages;
     }
